解决啦哈哈哈，因为PID里面Pout大小超过变量定义极限:(

+++ SB_Down/Chassis/USER/DriverLayer/user_pid.c
@@ -1,6 +1,6 @@
 #include "user_pid.h"
 
-int16_t limit_max(int16_t value,int16_t Max_out)
+int16_t limit_max(int32_t value,int32_t Max_out)
 {
 	if(value > Max_out){
 		value = Max_out;
